refactored components in human feedback modal
